compatibility assistance for manually edited autosar version headers
The naming scheme for the autosar xsd files changed after version 4.3.0.
If someone edits the version in an old file and blindly extends it to autosar 4.3.1 or newer, then it becomes invalid.
Specifically, Autosar 4.3.0 uses the xsd file AUTOSAR_4-3-0.xsd, but Autosar 4.3.1 uses AUTOSAR_00044.xsd.
Plausible invalid headers for 4.3.1, 4.4.0 and 4.5.0 can now be recognized and corrected.