Improved reference handling
The previous code which requred the content of the DEST attribute of a
reference to be equal to the referenced element name was too strict.
A field had to be added to the specification which states what the
valid value for DEST is in a reference referring to that element.
This new information is now used by get/set_reference_target.