Stop locking dependency versions
Thiserror was locked to 2.0.x and smallvec to 1.13.x. From the history
and some testing this doesn't seem to be required, so let's stop to
allow updates in downstream dependencies.

This is not yet relevant for thiserror, as it is at 2.0.11, but it is
relevant for smallvec which has released 1.14.x.